Moved king move rules into the move module.
